+++ src/Exception/NotDirectoryException.php
@@ -17,6 +17,7 @@ declare(strict_types=1);
 namespace Tools\Exception;
 
 use Tools\FileException;
+use Tools\Filesystem;
 
 /**
  * "Filename is not a directory" exception.
@@ -33,7 +34,7 @@ class NotDirectoryException extends FileException
     public function __construct(?string $message = null, int $code = 0, ?\Throwable $previous = null, ?string $path = null)
     {
         if (!$message) {
-            $message = $path ? sprintf('Filename `%s` is not a directory', rtr($path)) : 'Filename is not a directory';
+            $message = $path ? sprintf('Filename `%s` is not a directory', (new Filesystem())->rtr($path)) : 'Filename is not a directory';
         }
         parent::__construct($message, $code, $previous, $path);
     }