+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # 1.x branch
 ## 1.4 branch
+### 1.4.4
+* added `Filesystem` class, all filesystem global function are now deprecated;
+* added `Filesystem::getRoot()` method;
+* added `\Tools\TestSuite\BackwardCompatibilityTrait` to provide methods to
+    achieve PHPUnit backward compatibility;
+* `TestTrait::assertFilePerms()` is deprecated. Use instead
+    `assertFileIsReadable()`/`assertFileIsWritable()`/
+    `assertDirectoryIsReadable()`/`assertDirectoryIsWritable()`;
+* `fileperms_as_octal()` and `fileperms_to_string()` global functions are now
+    deprecated and will be removed in a future release;
+* some functions have been moved to `array_functions.php` file.
+
 ### 1.4.3
 * added `array_unique_recursive()` global function;
 * added `Exceptionist::methodExists()` method;